+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.4.1.0
+=======
+
+* Add `AuthCheckPermissionsMethod` endpoint.
+* Add `AuthGetUserIdMethod` endpoint.
+
 0.4.0.1
 =======